\name{ERN_OSM}
\alias{ERN_OSM}
\alias{ERN_OSM_correct}
\alias{ERN_OSM.nt}
\alias{ERN_OSM_cor.nt}
\docType{data}
\title{Estevan road network from OpenStreetMap (SpatialLinesDataFrame)}
\description{
  This data set is a part of Estevan road network downloaded from OpenStreetMap, 
  and a version with topology errors corrected is also incorporated.
}
\usage{
data(ERN_OSM)
data(ERN_OSM_correct)
}
\format{
 Two \dQuote{SpatialLinesDataFrame} objects named \dQuote{ERN_OSM.nt} and \dQuote{ERN_OSM_cor.nt}.
  }
\source{
  \url{https://www.openstreetmap.org/#map=14/49.1388/-102.9961}
}
\references{
  OpenStreetMap. 2014. Estevan road network. ed. O. community.
}
