\name{Nodes.coordinates}
\alias{Nodes.coordinates}
\title{Return the coordinates of the given \dQuote{nodelist}}
\description{
The function returns the coordinates of the given \dQuote{nodelist} in two 
columns X and Y.
}
\usage{
Nodes.coordinates(nodelist)
}
\arguments{
 \item{nodelist}{An \dQuote{nodelist} object}
}
\value{
  \item{Nodesxy}{A matrix with two columns X and Y} 
}
\author{Binbin Lu \email{lubinbin220@gmail.com}}
\keyword{coordinate, node}
