\name{ORN}
\alias{ORN}
\alias{rn}
\docType{data}
\title{Ontario road network data (SpatialLinesDataFrame)}
\description{
  This data set is a part of Ontario road network data set downloaded from 
  Ontario Ministry of Natural Resources.
}
\usage{data(ORN)}
\format{
 A \dQuote{SpatialLinesDataFrame} object.
  }
\source{
  \url{http://www.geographynetwork.ca/website/orn/viewer.htm}
}
\references{
  Peterborough, ON: Ontario Ministry of Natural Resources, Ontario Road Network. 
  Available: \url{http://www.geographynetwork.ca/website/orn/viewer.htm}, 2006, 
  (Accessed Dec. 4, 2009)
}
\keyword{road network data}
