\name{Nodes.coordinates}
\alias{Nodes.coordinates}
\title{Return the coordinates of nodes in a \dQuote{\link{nodelist}}}
\description{
A function that returns the coordinates of nodes in a two-column matrix (X,Y) from
a \dQuote{\link{nodelist}}
}
\usage{
Nodes.coordinates(nodelist)
}
\arguments{
 \item{nodelist}{A \dQuote{\link{nodelist}} object}
}
\value{
  \item{Nodesxy}{A two-column matrix (X, Y)} 
}
\author{Binbin Lu \email{lubinbin220@gmail.com}}
\keyword{coordinate, node}
