\name{Nodes.coordinates}
\title{Output the coordinates of the given nodelist}
\description{
The function returns the coordinates of the given nodelist in two columns X and Y.
}
\usage{
Nodes.coordinates(nodelist)
}
\arguments{
 \item{nodelist}{An nodelist object, see \link{nodelist}}
}
\value{
  \item{Nodesxy}{A matrix with two columns X and Y} 
}
\author{Binbin Lu \email{lubinbin220@gmail.com}}
\keyword{coordinate, node}
